# Conflicts:
#	src/backend/ci/core/quality/biz-quality/src/main/kotlin/com/tencent/devops/quality/dao/v2/QualityIndicatorDao.kt

+++ src/backend/ci/core/process/biz-process/src/main/kotlin/com/tencent/devops/process/engine/atom/vm/DispatchVMStartupTaskAtom.kt
@@ -133,7 +133,7 @@ class DispatchVMStartupTaskAtom @Autowired constructor(
         val pipelineInfo = pipelineRepositoryService.getPipelineInfo(projectId, pipelineId)
             ?: throw BuildTaskException(
                 errorType = ErrorType.SYSTEM,
-                errorCode = ERROR_PIPELINE_NOT_EXISTS,
+                errorCode = ERROR_PIPELINE_NOT_EXISTS.toInt(),
                 errorMsg = "流水线不存在",
                 pipelineId = pipelineId,
                 buildId = buildId,
@@ -143,7 +143,7 @@ class DispatchVMStartupTaskAtom @Autowired constructor(
         val model = pipelineBuildDetailService.getBuildModel(buildId)
             ?: throw BuildTaskException(
                 errorType = ErrorType.SYSTEM,
-                errorCode = ERROR_PIPELINE_MODEL_NOT_EXISTS,
+                errorCode = ERROR_PIPELINE_MODEL_NOT_EXISTS.toInt(),
                 errorMsg = "流水线模型不存在",
                 pipelineId = pipelineId,
                 buildId = buildId,
@@ -153,7 +153,7 @@ class DispatchVMStartupTaskAtom @Autowired constructor(
         val container = model.getContainer(vmSeqId)
             ?: throw BuildTaskException(
                 errorType = ErrorType.SYSTEM,
-                errorCode = ERROR_PIPELINE_NODEL_CONTAINER_NOT_EXISTS,
+                errorCode = ERROR_PIPELINE_NODEL_CONTAINER_NOT_EXISTS.toInt(),
                 errorMsg = "流水线的模型中指定构建容器${vmNames}不存在",
                 pipelineId = pipelineId,
                 buildId = buildId,
@@ -349,7 +349,7 @@ class DispatchVMStartupTaskAtom @Autowired constructor(
     private fun agentException(pipelineId: String, buildId: String) =
         BuildTaskException(
             errorType = ErrorType.SYSTEM,
-            errorCode = ERROR_PIPELINE_AGENT_STATUS_EXCEPTION,
+            errorCode = ERROR_PIPELINE_AGENT_STATUS_EXCEPTION.toInt(),
             errorMsg = "第三方构建机的Agent状态异常",
             pipelineId = pipelineId,
             buildId = buildId,
